requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai # or whatever LLM client you're using
